Add unverified message to profile page Cleanup alert API to allow easier compositions for unverified profile pag Cleanup main nav dropdown links Move alert icons to main icon folder
