+++ src/components/Layout/Header/UserSection.js
@@ -99,7 +99,7 @@ const UserSection = ({ links, setDropdownOpen, isDropdownOpen }) => {
                 className='w-full text-left px-4 py-2 text-sm leading-5 text-gray-700 hover:bg-gray-100 focus:outline-none focus:bg-gray-100 transition duration-150 ease-in-out'
                 role='menuitem'
               >
-                log out
+                Log out
               </button>
             </div>
           </div>
@@ -112,13 +112,13 @@ const UserSection = ({ links, setDropdownOpen, isDropdownOpen }) => {
         to='/login'
         className='mr-3 inline-flex items-center px-3 py-2 border border-gray-300 shadow-sm text-sm leading-4 font-medium rounded-md text-gray-700 bg-white hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500'
       >
-        log in
+        Log in
       </Link>
       <Link
         to='/create-account'
         className='inline-flex items-center px-3 py-2 border border-transparent text-sm leading-4 font-medium rounded-md shadow-sm text-white bg-blue-600 hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500'
       >
-        create account
+        Create account
       </Link>
     </div>
   )
