+++ src/helper/auth.js
@@ -1,5 +1,32 @@
 import axios from 'axios'
-import { AUTH_API } from 'config'
+import { AUTH_API, VERIFY_API } from 'config'
+import { alertType } from 'context/AlertContext'
+
+export const createUnverifiedAlert = (email) => ({
+  type: alertType.INFO,
+  header: 'Your account is unverified',
+  message:
+    'Currently you cannot access all features due to being unverified. Please check your email for a verification link. You may also click the "Resend email" button below if you cannot find your email.',
+  close: true,
+  action: {
+    onClick: async (success, fail, load) => {
+      try {
+        load()
+        await axios.post(
+          VERIFY_API + '/resend',
+          { email },
+          { withCredentials: true }
+        )
+        success()
+      } catch (e) {
+        fail()
+      }
+    },
+    buttonText: 'Resend email',
+    successMessage: 'Email sent!',
+    failMessage: 'Sending email failed. Please try again later.',
+  },
+})
 
 export const logoutAPI = async () => {
   // This is to support logging out from all windows
